refactor(agent): split api_server/receipts.rs into receipts/{emit,verify,...}
Decomposes the 3,827-line `api_server/receipts.rs` god file into a
`receipts/` directory with 25 focused submodules (mod.rs + 24
sub-files). The largest new file is `network_extension.rs` at 556
lines; every other file is under 415 lines. All `pub(crate)` items
remain `pub(crate)` and are re-exported from `receipts/mod.rs` so
`api_server.rs`'s `pub(crate) use receipts::*;` continues to expose
them unchanged. No function signatures, types, or call sites were
touched outside the receipts module.

Module map:

- emit_detection, emit_policy, emit_lifecycle, emit_deception,
  emit_response — async receipt emission helpers, grouped by event
  family.
- filter, accessors, severity — predicate + accessor helpers over
  signed-receipt `endpointDecision` metadata.
- verify_execution, verify_lifecycle, verify_effects, verify_rollback,
  verify_actor, verify_lifecycle_lookup, verify_signature — pure
  proof-contract verifiers for response execution receipts.
- provider_evidence, network_extension — sensor-state and per-provider
  evidence builders.
- control_store — wire conversion from signed endpoint receipts to
  control-API upload payloads.
- policy_snapshot, raw_artifact_approval — local policy YAML parsing
  and privacy/approval gating.
- actor, signer — endpoint actor derivation and signer key loading.
- ledger_index — JSONL receipt ledger persistence and sidecar index
  file maintenance.

Submodules pull in api_server scope via `use super::super::*;`,
mirroring the original `use super::*;` pattern used by sibling
`api_server/<name>.rs` files. Build and test counts (205 passed / 26
pre-existing failures) match the baseline; no new failures.
